pyproject.toml
[project]
name = "reading-list-manager"
version = "0.1.0"
description = ""
authors = [
    { name = "Vitalii Derda", email = "derdavitaliy@gmail.com" }
]
license = { text = "GNU GENERAL PUBLIC LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular-sidecar (>=2025.2.1,<2026.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
